Allow multiple diagrams to be opened (#44)
* Allow multiple diagrams to be opened

- Fix TheiaJsonrpcGLSPClient to only have one handler for connection
- Fix problem with filter with same ID being added multiple times

- Add formatting of CSS files to prettier
- Centralize ID handling for containers and global IDs
-- Diagram elements need locally unique IDs
-- Other elements need globally unique IDs

Fixes https://github.com/CrossBreezeNL/crossmodel/issues/16

* Updated playwright config to have timeouts for action and navigation so tests fail on actions i.s.o. timeout.

* Added logging of the id of the node it couldn't find.

* Fix "Couldn't find node" when selecting something other than an entity

* Fix "Property widget starts flashing"

Problem: We open an entity through the property view. That triggers a
rebuild of the entity document which triggers an update in all open
diagrams that have that entity as they have a reference to it. The
diagram update also resets the selection state in each diagram which
in turn updates the property view again as it reacts to selection
changes. This can lead to an update-cycle if there is more than one
diagram open with the same entity.

Fix: Since the property view reacts to selection changes and updates in
the diagram trigger such a change, there is no need for the property
view to listen actually participate in the model lifecycle itself. So
there is no need to do the open/close since we do not need to listen to
updates ourselves.

* Fix Diagram context menu issue

- Ensure that the navigator tree always gets a URI of the selected node

* Improve update handling between editors

- Ensure GLSP selection is forwarded again if active editor changes
- Only update properties if selection of active widget changes
- Add additional change reason for better update decision in GLSP

* Added example relationship and diagram to e2e-tests example worspace.
Added context menu tests on relationship and diagram files in the explorer view.

---------

Co-authored-by: Harmen Wessels <97173058+harmen-xb@users.noreply.github.com>